refactor: 클래스 위치 변경 및 불필요한 import 삭제
