+++ src/main/java/blackjack/domain/CardDeck.java
@@ -1,5 +1,6 @@
-package blackjack.domain.card;
+package blackjack.domain;
 
+import blackjack.domain.card.Card;
 import java.util.LinkedList;
 import java.util.Queue;
