+++ src/test/java/blackjack/domain/CardDeckGeneratorTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
